requirements.txt
dbt-core
dbt-postgres
nest_asyncio
uvicorn
telethon
psycopg2-binary
sqlalchemy
pandas
python-dotenv
dbt-postgres
ultralytics
fastapi
dagster
dagster-webserver
